Migrated OCPI Command endpoint
